Add install check process

+++ download.sh
@@ -8,27 +8,35 @@ cat << EOS
 
 EOS
 
+function command_exists {
+  command -v "$1" > /dev/null;
+}
+
 #
 # Install homebrew.
 #
-echo " --------- Homebrew ----------"
-/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
-brew update
-brew upgrade --all --cleanup
-brew -v
-echo " ------------ END ------------"
+if ! command_exists brew ; then
+  echo " --------- Homebrew ----------"
+  /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+  brew update
+  brew upgrade --all --cleanup
+  brew -v
+  echo " ------------ END ------------"
+fi
 
 #
 # Install git
 #
-echo " ------------ Git ------------"
-brew install git
-git --version
-echo " ------------ END ------------"
+if ! command_exists git ; then
+  echo " ------------ Git ------------"
+  brew install git
+  git --version
+  echo " ------------ END ------------"
+fi
 
 #
-# noumi0k/mac-auto-setup.git
+# mac-auto-setup.git
 #
-echo " ------------ noumi0k/mac-auto-setup.git ------------"
+echo " ---- mac-auto-setup.git -----"
 git clone https://github.com/AkkeyLab/mac-auto-setup.git
 echo " ------------ END ------------"
