+++ setup.sh
@@ -8,6 +8,10 @@ cat << EOS
 
 EOS
 
+function command_exists {
+  command -v "$1" > /dev/null;
+}
+
 #
 # Memorize user pass
 #
@@ -16,27 +20,33 @@ read -sp "Your Password: " pass;
 #
 # Mac App Store apps install
 #
-echo " ---- Mac App Store apps -----"
-brew install mas
-mas install 497799835  # Xcode (8.2.1)
-echo " ------------ END ------------"
+if ! command_exists mas ; then
+  echo " ---- Mac App Store apps -----"
+  brew install mas
+  mas install 497799835  # Xcode (8.2.1)
+  echo " ------------ END ------------"
+fi
 
 #
 # Install zsh
 #
-echo " ------------ zsh ------------"
-brew install zsh zsh-autosuggestions zsh-completions zsh-syntax-highlighting colordiff
-which -a zsh
-echo $pass | sudo -S -- sh -c 'echo '/usr/local/bin/zsh' >> /etc/shells'
-chsh -s /usr/local/bin/zsh
-echo " ------------ END ------------"
+if ! command_exists zsh ; then
+  echo " ------------ zsh ------------"
+  brew install zsh zsh-autosuggestions zsh-completions zsh-syntax-highlighting colordiff
+  which -a zsh
+  echo $pass | sudo -S -- sh -c 'echo '/usr/local/bin/zsh' >> /etc/shells'
+  chsh -s /usr/local/bin/zsh
+  echo " ------------ END ------------"
+fi
 
 #
 # Install vim
 #
-echo " ------------ Vim ------------"
-brew install vim --with-override-system-vi
-echo " ------------ END ------------"
+if ! command_exists vim ; then
+  echo " ------------ Vim ------------"
+  brew install vim --with-override-system-vi
+  echo " ------------ END ------------"
+fi
 
 #
 # Powerline
@@ -51,17 +61,19 @@ echo " ------------ END ------------"
 #
 # Install ruby
 #
-echo " ----------- Ruby ------------"
-brew install rbenv
-brew install ruby-build
-rbenv --version
-rbenv install -l
-ruby_latest=$(rbenv install -l | grep -v '[a-z]' | tail -1 | sed 's/ //g')
-rbenv install $ruby_latest
-rbenv global $ruby_latest
-rbenv rehash
-ruby -v
-echo " ------------ END ------------"
+if ! command_exists rbenv ; then
+  echo " ----------- Ruby ------------"
+  brew install rbenv
+  brew install ruby-build
+  rbenv --version
+  rbenv install -l
+  ruby_latest=$(rbenv install -l | grep -v '[a-z]' | tail -1 | sed 's/ //g')
+  rbenv install $ruby_latest
+  rbenv global $ruby_latest
+  rbenv rehash
+  ruby -v
+  echo " ------------ END ------------"
+fi
 
 #
 # Install dotfiles system
@@ -75,62 +87,74 @@ echo " ------------ END ------------"
 #
 # Install Node.js env
 #
-echo " ---------- Node.js ----------"
-curl -L git.io/nodebrew | perl - setup
-nodebrew ls-remote
-nodebrew install-binary latest
-nodebrew ls
-nodebrew use latest
-node -v
-npm -v
-echo " ------------ END ------------"
+if ! command_exists nodebrew ; then
+  echo " ---------- Node.js ----------"
+  curl -L git.io/nodebrew | perl - setup
+  nodebrew ls-remote
+  nodebrew install-binary latest
+  nodebrew ls
+  nodebrew use latest
+  node -v
+  npm -v
+  echo " ------------ END ------------"
+fi
 
 #
 # Install Yarn
 #
-echo " ----------- Yarn ------------"
-brew install yarn
-echo " ------------ END ------------"
+if ! command_exists yarn ; then
+  echo " ----------- Yarn ------------"
+  brew install yarn
+  echo " ------------ END ------------"
+fi
 
 #
 # TeX settings
 #
-echo " ------------ TeX ------------"
-brew cask install mactex
-# Tex Live Utility > preference > path -> /Library/TeX/texbin
-version=$(tex -version | grep -oE '2[0-9]{3}' | head -1)
-echo $pass | sudo -S /usr/local/texlive/$version/bin/x86_64-darwin/tlmgr path add
-echo $pass | sudo -S tlmgr update --self --all
-# JPN Lang settings
-cd /usr/local/texlive/$version/texmf-dist/scripts/cjk-gs-integrate
-echo $pass | sudo -S perl cjk-gs-integrate.pl --link-texmf --force
-echo $pass | sudo -S mktexlsr
-echo $pass | sudo -S kanji-config-updmap-sys hiragino-elcapitan-pron
-# Select ==> TeXShop > Preferences > Source > pTeX (ptex2pdf)
-echo " ------------ END ------------"
+if ! command_exists tex ; then
+  echo " ------------ TeX ------------"
+  brew cask install mactex
+  # Tex Live Utility > preference > path -> /Library/TeX/texbin
+  version=$(tex -version | grep -oE '2[0-9]{3}' | head -1)
+  echo $pass | sudo -S /usr/local/texlive/$version/bin/x86_64-darwin/tlmgr path add
+  echo $pass | sudo -S tlmgr update --self --all
+  # JPN Lang settings
+  cd /usr/local/texlive/$version/texmf-dist/scripts/cjk-gs-integrate
+  echo $pass | sudo -S perl cjk-gs-integrate.pl --link-texmf --force
+  echo $pass | sudo -S mktexlsr
+  echo $pass | sudo -S kanji-config-updmap-sys hiragino-elcapitan-pron
+  # Select ==> TeXShop > Preferences > Source > pTeX (ptex2pdf)
+  echo " ------------ END ------------"
+fi
 
 #
 # Install wget
 #
-echo " ----------- wget ------------"
-brew install wget
-wget --version
-echo " ------------ END ------------"
+if ! command_exists wget ; then
+  echo " ----------- wget ------------"
+  brew install wget
+  wget --version
+  echo " ------------ END ------------"
+fi
 
 #
 # CocoaPods
 #
-echo " --------- CocoaPods ---------"
-echo $pass | sudo -S gem install -n /usr/local/bin cocoapods --pre
-pod setup
-echo " ------------ END ------------"
+if ! command_exists pod ; then
+  echo " --------- CocoaPods ---------"
+  echo $pass | sudo -S gem install -n /usr/local/bin cocoapods --pre
+  pod setup
+  echo " ------------ END ------------"
+fi
 
 #
 # Carthage
 #
-echo " --------- Carthage ----------"
-brew install carthage
-echo " ------------ END ------------"
+if ! command_exists carthage ; then
+  echo " --------- Carthage ----------"
+  brew install carthage
+  echo " ------------ END ------------"
+fi
 
 while true; do
   read -p 'Now install web apps? [Y/n]' Answer
